plat: marvell: Use 'int' type for fields in sata_cfg structure
struct sata_cfg saves qlm to gser mapping information in to_gser field
and qlm to lane mapping in to_lane field. The function that populates
these fields return 'int' type.

Reviewed-on: https://sj1git1.cavium.com/3050
Tested-by: sa_ip-sw-jenkins

+++ include/plat/marvell/octeontx/octeontx_scfg.h
@@ -11,8 +11,8 @@
 #include <platform_scfg.h>
 
 typedef struct sata_cfg {
-	uint8_t to_gser[MAX_SATA];
-	uint8_t to_lane[MAX_SATA];
+	int to_gser[MAX_SATA];
+	int to_lane[MAX_SATA];
 } sata_cfg_t;
 
 typedef struct scfg {
